Upt. 설명 주석 추가
