+++ lib/global/components/ds_calendar.dart
@@ -3,14 +3,15 @@ import 'dart:io';
 
 import 'package:intl/intl.dart';
 import 'package:my_workout_diary_app/global/style/ds_colors.dart';
+import 'package:my_workout_diary_app/pages/02_record/page_record.dart';
 import 'package:table_calendar/table_calendar.dart';
 import 'package:flutter/material.dart';
 
-TableCalendar DSTableCalendar({
-  required DateTime? selectedDay,
-  required void Function(DateTime, DateTime)? onDaySelected,
-  required DateTime focusedDay,
-}) {
+TableCalendar DSTableCalendar(
+    {required DateTime? selectedDay,
+    required void Function(DateTime, DateTime)? onDaySelected,
+    required DateTime focusedDay,
+    required List<Event> Function(DateTime)? eventLoader}) {
   return TableCalendar<Event>(
     selectedDayPredicate: (day) => isSameDay(selectedDay, day),
     onDaySelected: onDaySelected,
@@ -24,7 +25,7 @@ TableCalendar DSTableCalendar({
     availableCalendarFormats: {
       CalendarFormat.month: CalendarFormat.month.name,
     },
-    eventLoader: _getEventsForDay,
+    eventLoader: eventLoader,
     headerStyle: const HeaderStyle(
       titleTextStyle: TextStyle(color: DSColors.gray6, fontSize: 16, fontWeight: FontWeight.w500),
       formatButtonTextStyle: TextStyle(color: DSColors.gray6),
@@ -156,42 +157,4 @@ Widget? dsMarker(int count) {
           ),
         )
       : null;
-}
-
-// test data
-final kEvents = LinkedHashMap<DateTime, List<Event>>(
-  equals: isSameDay,
-  hashCode: getHashCode,
-)..addAll(_kEventSource);
-
-final _kEventSource = Map.fromIterable(List.generate(50, (index) => index),
-    key: (item) => DateTime.utc(kFirstDay.year, kFirstDay.month, item * 5),
-    value: (item) => List.generate(item % 4 + 1, (index) => Event('Event $item | ${index + 1}')))
-  ..addAll({
-    kToday: [
-      Event('Today\'s Event 1'),
-      Event('Today\'s Event 2'),
-    ],
-  });
-
-int getHashCode(DateTime key) {
-  return key.day * 1000000 + key.month * 10000 + key.year;
-}
-
-final kToday = DateTime.now();
-final kFirstDay = DateTime(kToday.year, kToday.month - 3, kToday.day);
-final kLastDay = DateTime(kToday.year, kToday.month + 3, kToday.day);
-
-class Event {
-  final String title;
-
-  const Event(this.title);
-
-  @override
-  String toString() => title;
-}
-
-List<Event> _getEventsForDay(DateTime day) {
-  // Implementation example
-  return kEvents[day] ?? [];
 }
